+++ backend/common/enums/gvdbs_types.py
@@ -1,4 +1,8 @@
+from typing import Literal
+
 class GVDBsTypes:
     CHROMA = 'chroma'
     QDRANT = 'qdrant'
     PGVECTOR = 'pgvector'
+
+type_GVDBsTypes = Literal['chroma', 'qdrant', 'pgvector']