+++ frontend/src/components/TopNavBar/TopNavBar.tsx
@@ -73,6 +73,9 @@ const TopNavBar = () => {
                   <LinkContainer to="/app/groupadmin/retrieval_parameters">
                     <Dropdown.Item>Retrieval Parameters</Dropdown.Item>
                   </LinkContainer>
+                  <LinkContainer to="/app/groupadmin/manage_vdbs">
+                    <Dropdown.Item>VDBs</Dropdown.Item>
+                  </LinkContainer>
                 </Dropdown.Menu>
               </Dropdown>
             )}